+++ logic/controller/adsrecommender_controler.py
@@ -16,7 +16,16 @@ class AdsRecommender(Recommender):
         self.ads = []
         self.mongo_helper = MongoHelper(Configs.mongo_url)
         self.neo4j_helper = Neo4jHelper(Configs.neo4j_url, Configs.neo4j_user, Configs.neo4j_pass)
-        
+
+    def _extract_context(self, subgraph):
+        contexts = []
+        for path in subgraph['paths']:
+            for edge in path['path']:
+                properties = json.loads(edge['properties'])
+                interpretation = properties['interpretation']
+                contexts.append(interpretation)
+        final_context = " ".join(contexts)
+        return final_context
 
     def _get_ads(self):
         campaigns = self.mongo_helper.get_all('campaigns')
